@@ -20,6 +20,7 @@
   "homepage": "https://github.com/nshew/typescript-sass-svg#README.md",
   "dependencies": {},
   "devDependencies": {
+    "@types/jquery": "^3.5.4",
     "clean-webpack-plugin": "^3.0.0",
     "css-loader": "^3.0.0",
     "file-loader": "^6.1.0",
@@ -30,7 +31,7 @@
     "style-loader": "^0.23.1",
     "svg-inline-loader": "^0.8.2",
     "ts-loader": "^6.0.3",
-    "typescript": "3.x",
+    "typescript": "4.x",
     "webpack": "^4.33.0",
     "webpack-cli": "^3.3.4",
     "webpack-dev-server": "^3.11.0"
