@@ -2,25 +2,22 @@
  * Since there is no JavaScript needed by the animation,
  * this file exists just to pull in the assets.
  */
-import './template.scss';
-const TemplateSVG = require('./template.svg');
+import './main.scss';
+// TODO: figure out why baseUrl imports aren't working
+import { MyApi } from './api';
 
-// add SVG object to body
-const svgEl = document.createElement('object');
-svgEl.setAttribute('type', 'image/svg+xml');
-svgEl.innerHTML = TemplateSVG;
-document.body.appendChild(svgEl);
+$((event) => {
+    const outputJQO = $('pre#output');
+    const myApi = new MyApi();
 
-// interface ICoords {
-//     x: number;
-//     y: number;
-// }
+    // myApi.getPledges({
+    //     dateBegin: '2020-07-07',
+    // }).done((result) => {
+    //     outputJQO.html(JSON.stringify(result, null, 2));
+    // });
 
-// window.findArcCoords = function (cx: number, cy: number, r: number, degrees=30): ICoords {
-//     const radians = degrees * Math.PI/180;
-
-//     const x = cx + r * Math.cos(radians);
-//     const y = cy - r * Math.sin(radians); // origin is upper-left, so subtract
-
-//     return { x, y };
-// };
+    myApi.getEvents({
+    }).done((result) => {
+        outputJQO.html(JSON.stringify(result, null, 2));
+    });
+});
