@@ -2,25 +2,27 @@ const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 
+const basePath = path.resolve(__dirname);
+
 module.exports = {
     mode:    'development',
     entry: {
-        app: path.resolve(__dirname, 'src/index.ts'),
+        app: basePath + '/src/index.ts',
     },
     plugins: [
         new CleanWebpackPlugin({ cleanStaleWebpackAssets: false }),
         new HtmlWebpackPlugin({
-            // TODO: customize https://github.com/jantimon/html-webpack-plugin
-            title: 'Template: TypeScript + Sass + SVG (HtmlWebpackPlugin)',
+            title: 'Givers & Gamers DEV',
+            template: basePath + '/src/index.html',
         }),
-    ], 
+    ],
     output:  {
-        path:     path.resolve(__dirname, 'dist'),
+        path:     basePath + '/dist',
         filename: '[name].bundle.js'
     },
     devtool: 'source-map',
     devServer: {
-        contentBase: './dist',
+        contentBase: basePath + '/dist',
     },
     resolve: {
         // for imports with no extension, resolve in this order
@@ -47,21 +49,21 @@ module.exports = {
                     'sass-loader'   // compiles Sass to CSS, using Node Sass by default
                 ]
             },
-            {
-                test: /\.(png|jpg|gif)$/,
-                exclude: /node_modules/,
-                use: [
-                  'file-loader',
-                ],
-            }, 
-            // TODO: Get SVGs in the bundle
-            {
-                test: /\.svg$/,
-                exclude: /node_modules/,
-                use: [
-                  'svg-inline-loader',
-                ],
-            },
+//            {
+//                test: /\.(png|jpg|gif)$/,
+//                exclude: /node_modules/,
+//                use: [
+//                  'file-loader',
+//                ],
+//            },
+//            // TODO: Get SVGs in the bundle
+//            {
+//                test: /\.svg$/,
+//                exclude: /node_modules/,
+//                use: [
+//                  'svg-inline-loader',
+//                ],
+//            },
         ]
     }
 };
